Update dependencies
So updating cypress is not an issue, but migrating to `intercept` won’t be as straight forward.
I think in the API `GET`s we’ll probably be able to do a fairly easy.. although slightly more complicated replacement.. and I don’t yet see an improvement.

However for our post/patch situations where `intercept` is particularly important, I haven’t yet found a one-size-fits-all solution.

A relatively one-to-one swap for what we currently test for would be something like:
```
Cypress.Commands.add('routeWithId', ({ method, url, response, staticResponse }) => {
  return cy.intercept(method, url, req => {
    const body = isFunction(response) ? response.call(this) : response;
    body.data.id = JSON.parse(req.body).data.id;
    req.reply(extend({ body }, staticResponse));
  },
  );
});
```

But there are so many variations, that perhaps just using `intercept` directly and manually updating the id is always better..  so.. still investigating, but we might as well update the dependenies.

# Conflicts:
#	package-lock.json
#	package.json