Switch releases to tags from branches
Unfortunately I believe the reason we didn’t use tags in the first place was that composer requires semver-like tag naming.  So custom tags and cool tower `/` nesting is not an option with tags.

That said, we don’t really have to manage releases that frequently, nor did we use custom release branches.  The prod/dev release system is pretty much all that’s needed it seems.  I think this will likely be sufficient.

When we update for the next release we’ll remove the `dev-` prefix in the BE `composer.json` and use the tag explicitly.

[sc-30255]