Ensure clinician/me stub aligns with groups/clinicians bootstrap
Previously it appears that Cypress consistently loaded the `/me` API prior to the groups bootstrap which allowed for tests to set the bootstrap and expect the correct data to be on the current clinician.  In v10, either might run first, breaking tests when the order changed.

This work aligns the stubs so that either load order will work.
[sc-30333]