Change directories to use entity
The API here is both singular and plural, but the type is plural, so we should rename it here as well.

The API also uses the slug over the id for reading.  As these are currently read-only there’s no current issue.

This PR should also temporarily drop the coverage until the epic is completed.