Remove changed to the picklist component
Changing the picklist class options broke possibly a poorly named `ui` attribute on the options list.

Perhaps we shouldn’t be using droplist for the main menu since it is so different than the standard

However,
I was able to reuse the `headingText` and `infoText` properties passed to the template to handle the workspace name and help url.
It’s a little dirty, but not crazy dirty and prevents needing to switch from droplist.

This also limits the awkward changes to the custom template.
Fixes [sc-34078]