Fix history goback
Newest iteration was keeping great track of history, but it wasn’t removing the current route so it would “go back” to itself.

I also added an `handleEmptyHistory` optional handler for if there’s no previous history.  This fixes @woody’s concern with direct links to forms not going back to anywhere useful.