Utilize history handleEmptyHistory for default route
This change looks bigger than it is.  The additional nested function was making the nested success handler in the `listenTo` rather deep, so I used `bindEvents` to make the service handlers more similar to `stateEvents`

This change looks bigger than it is.  The additional nested function was making the nested success handler in the `listenTo` rather deep, so I used `bindEvents` to make the service handlers more similar to `stateEvents`

This change looks bigger than it is.  The additional nested function was making the nested success handler in the `listenTo` rather deep, so I used `bindEvents` to make the service handlers more similar to `stateEvents`