Move disabled checks after navigate
I’m not 100% sure why this works.
The first bootstrap likely failed occassionally due to the bootstrap waits, but the second `navigate` works where as `go(‘forward’)` is flaky.  My best guess… I think `go` resolves after it realizes the hash changes and not the whole page url changes which likely happens _after_ the render, where-as `navigate` immediately moves on to the next test.

I’m not 100% sure why this works.
The first bootstrap likely failed occassionally due to the bootstrap waits, but the second `navigate` works where as `go(‘forward’)` is flaky.  My best guess… I think `go` resolves after it realizes the hash changes and not the whole page url changes which likely happens _after_ the render, where-as `navigate` immediately moves on to the next test.