Add guard to prevent extra requests in form test definition
On firefox these tests were failing due to `postMessage` not being able to serialize the submission.

https://cloud.cypress.io/projects/ep9zr6/runs/4e2c7d3d-74fe-4daf-a42e-7818d1088058/test-results/b45a527c-68fe-4f4d-b59f-ee588d331bf6/screenshots

I don’t really know what is going on, but I do know that form.io hammers these end points.  This keeps it to a single request which I think will prevent the firefox failure.

Notably I could not get this to fail in Firefox locally.