Filter out 401/500 errors
These errors when not handled (which they generally do not need to be) throw `Uncaught [Response]`

Currently this is the only place I can find where we’re rejecting the promise specifically.  All other errors simply fall through and are handled or possibly ignored.

These errors do not need to be logged because the 401/500s are