Add cache buster to patient endpoint
[sc-52969]

This uses `filter[workspace]={id}`
For our usecase because workspace is in the header it doesn’t actually matter if this API is functional on it’s own for filtering.

But this will prevent cross workspace caching of this particular endpoint.  This ought to be enough even including other `/api/patient/{id}/relationships/foo` endpoints as at least in our current implementation, a base patient call is always made prior to the relationship ones.