Clear localstorage on a 5xx error
Current 5xx errors cannot be handled explicitly by an onFail.  I think this is still correct, but we can detect when a page is routing to a 5xx error.  In these cases we can clear the localstorage for the current state which should help possibly prevent a 500 doom loop assuming the 500 isn’t occurring on the default filter set.