WIP: Use default mutator for collection mocking
This probably can’t be implemented until all the merge json api refactoring is complete, but one thing I’ve noticed is a steady slowing down of tests on CI.

In particular the worklist tests are taking 11 minutes on CI.

I _think_ this is because often we’ll setup a generic stub of a collection, by completely replacing it, meaning it has to do a lot of setup that is unncessary for the routes for a test.

There’s almost certainly more filtering we can do as right now we’re always including all relationationships possible on collections, when they’re reduced on prod.

But this should prove it can be improved.